Renamed AccountDomainWrapper to Account and several other related classes - RA-833
Added message translations for manage encounter types - RA-833

Renamed a couple of poorly named classes - RA-751
