+++ api/src/main/java/org/openmrs/module/adminui/AdminUiConstants.java
@@ -9,17 +9,16 @@
  */
 package org.openmrs.module.adminui;
 
-public class AdminUiConstants {
-
-    public static final String DAEMON_USER_UUID = "A4F30A1B-5EB9-11DF-A648-37A07F9C90FB";
-
-    //from EmrApi module
-
-    public static final String ROLE_PREFIX_CAPABILITY = "Application Role: ";
-
-    public static final String ROLE_PREFIX_PRIVILEGE_LEVEL = "Privilege Level: ";
-
-    public static final String PRIVILEGE_LEVEL_FULL_ROLE = ROLE_PREFIX_PRIVILEGE_LEVEL + "Full";
-
-
+public class AdminUIConstants {
+	
+	public static final String DAEMON_USER_UUID = "A4F30A1B-5EB9-11DF-A648-37A07F9C90FB";
+	
+	//from EmrApi module
+	
+	public static final String ROLE_PREFIX_CAPABILITY = "Application Role: ";
+	
+	public static final String ROLE_PREFIX_PRIVILEGE_LEVEL = "Privilege Level: ";
+	
+	public static final String PRIVILEGE_LEVEL_FULL_ROLE = ROLE_PREFIX_PRIVILEGE_LEVEL + "Full";
+	
 }
