+++ api/src/main/java/org/openmrs/module/adminui/account/AccountService.java
@@ -9,49 +9,52 @@
  */
 package org.openmrs.module.adminui.account;
 
-import org.openmrs.Role;
-import org.openmrs.Person;
 import java.util.List;
 
+import org.openmrs.Person;
+import org.openmrs.Role;
+import org.openmrs.module.adminui.AdminUIConstants;
+
 public interface AccountService {
-
-	 /**
-     * Save the account details to the database
-     *
-     * @param account
-     * @return
-     */
-    void saveAccount(AccountDomainWrapper account);
-    
-    /**
-     * @return
-     * @should get all unique accounts
-     */
-    List<AccountDomainWrapper> getAllAccounts();
-    
-    /**
-     * Gets all Capabilities, i.e roles with the {@link org.openmrs.module.adminui.AdminUiConstants#ROLE_PREFIX_CAPABILITY} prefix
-     *
-     * @return a list of Roles
-     * @should return all roles with the capability prefix
-     */
-    List<Role> getAllCapabilities();
-
-    /**
-     * Gets all Privilege Levels, i.e roles with the
-     *
-     * @return a list of Roles
-     * @should return all roles with the privilege level prefix
-     */
-    List<Role> getAllPrivilegeLevels();
-    
-    /**
-     * Gets an account for the Specified person object
-     *
-     * @return
-     * @should return the account for the specified person if they are associated to a user
-     * @should return the account for the specified person if they are associated to a provider
-     */
-    AccountDomainWrapper getAccountByPerson(Person person);
-
+	
+	/**
+	 * Save the account details to the database
+	 *
+	 * @param account
+	 * @return
+	 */
+	void saveAccount(Account account);
+	
+	/**
+	 * @return
+	 * @should get all unique accounts
+	 */
+	List<Account> getAllAccounts();
+	
+	/**
+	 * Gets all Capabilities, i.e roles with the {@link AdminUIConstants#ROLE_PREFIX_CAPABILITY}
+	 * prefix
+	 *
+	 * @return a list of Roles
+	 * @should return all roles with the capability prefix
+	 */
+	List<Role> getAllCapabilities();
+	
+	/**
+	 * Gets all Privilege Levels, i.e roles with the
+	 *
+	 * @return a list of Roles
+	 * @should return all roles with the privilege level prefix
+	 */
+	List<Role> getAllPrivilegeLevels();
+	
+	/**
+	 * Gets an account for the Specified person object
+	 *
+	 * @return
+	 * @should return the account for the specified person if they are associated to a user
+	 * @should return the account for the specified person if they are associated to a provider
+	 */
+	Account getAccountByPerson(Person person);
+	
 }
