+++ api/src/main/java/org/openmrs/module/adminui/adminuiActivator.java
@@ -9,8 +9,7 @@
  */
 package org.openmrs.module.adminui;
 
-
-import org.apache.commons.logging.Log; 
+import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.openmrs.module.BaseModuleActivator;
 import org.openmrs.module.ModuleActivator;
@@ -18,10 +17,10 @@ import org.openmrs.module.ModuleActivator;
 /**
  * This class contains the logic that is run every time this module is either started or stopped.
  */
-public class adminuiActivator extends BaseModuleActivator {
+public class AdminUIActivator extends BaseModuleActivator {
 	
 	protected Log log = LogFactory.getLog(getClass());
-		
+	
 	/**
 	 * @see ModuleActivator#willRefreshContext()
 	 */
@@ -63,5 +62,5 @@ public class adminuiActivator extends BaseModuleActivator {
 	public void stopped() {
 		log.info("adminui Module stopped");
 	}
-		
+	
 }
