+++ omod/src/main/java/org/openmrs/module/adminui/page/controller/metadata/locations/LocationPageController.java
@@ -18,7 +18,6 @@ import org.openmrs.LocationAttributeType;
 import org.openmrs.LocationTag;
 import org.openmrs.api.LocationService;
 import org.openmrs.api.context.Context;
-import org.openmrs.module.adminui.AdminUiConstants;
 import org.openmrs.module.uicommons.UiCommonsConstants;
 import org.openmrs.ui.framework.annotation.BindParams;
 import org.openmrs.ui.framework.annotation.SpringBean;
